Move `PageFormatInfo` and related to `document.layout.page`
