+++ quarkdown-core/src/main/kotlin/com/quarkdown/core/ast/quarkdown/invisible/PageMarginContentInitializer.kt
@@ -2,7 +2,7 @@ package com.quarkdown.core.ast.quarkdown.invisible
 
 import com.quarkdown.core.ast.NestableNode
 import com.quarkdown.core.ast.Node
-import com.quarkdown.core.document.page.PageMarginPosition
+import com.quarkdown.core.document.layout.page.PageMarginPosition
 import com.quarkdown.core.visitor.node.NodeVisitor
 
 /**
